Created tests, added and edited documentation, and refactored some methods.
Created tests for testing drive PID (issue #3), drive odometry (issue #4), shooter hood (issue #6), shooter vision (issue #5), and shooter auto shoot (issue #7) via creating a testing framework that runs the tests in test mode and allows for the editing of values in the wpilib Shuffleboard. Added and edited documentation (issue #2) for the new tests and new methods that the tests required of other subsystems. Mainly refactored the setPIDF() method that was located in Robot.java, which has now been moved to DriveSystem.java in the form of two overloaded methods, one that accepts a drive mode enum to set the PIDF values and one that directly accepts PIDF values.