Added code for High Climb Arm Limit Switches.
Added code for the high climb arm limit switches (Fixes #11) to where if either of the limit switches are pressed, then the high climb arm motor will be stopped once and will not allow movement against the currently pressed limit switch. The documentation was altered in lieu of the changes.